requirements.txt
Adafruit-BluefruitLE==0.9.10
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==0.2.3
autopep8==1.4.3
btlewrap==0.0.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorclass==2.2.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pathlib==1.0.1
pbr==5.1.3
pip-check==2.3.3
PyBluez==0.22
pycodestyle==2.5.0
pyobjc-core==5.2
pyobjc-framework-Cocoa==5.2
pyserial==3.4
PyYAML==5.1
pyyml==0.0.2
requests==2.21.0
six==1.12.0
smars-library==0.1.3
SQLAlchemy==1.3.1
stevedore==1.30.1
terminaltables==3.1.0
urllib3==1.25
virtualenv==16.4.3
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
